requirements.txt
langchain==0.0.79
faiss-cpu==1.7.3
openai==0.27.2
tenacity==8.2.0
bs4
fastapi
uvicorn[standard]
python-dotenv